+++ kohii-core/src/main/java/kohii/v1/core/Playback.kt
@@ -122,7 +122,9 @@ abstract class Playback(
     val tokenUpdateListener: TokenUpdateListener? = null,
     val networkTypeChangeListener: NetworkTypeChangeListener? = null,
     val initialPlaybackInfo: PlaybackInfo? = null,
-    val callbacks: Set<Callback> = emptySet()
+    val callbacks: Set<Callback> = emptySet(),
+    val rendererAttachedCallback: RendererAvailabilityCallback? = null,
+    val rendererDetachedCallback: RendererAvailabilityCallback? = null
   )
 
   override fun toString(): String {
@@ -165,6 +167,8 @@ abstract class Playback(
   private var artworkHintListener: ArtworkHintListener? = null
   private var tokenUpdateListener: TokenUpdateListener? = null
   private var networkTypeChangeListener: NetworkTypeChangeListener? = null
+  private var rendererAttachedCallback: RendererAvailabilityCallback? = null
+  private var rendererDetachedCallback: RendererAvailabilityCallback? = null
 
   /**
    * Returns a usable renderer for the [playable], or `null` if no renderer is available or needed.
@@ -223,6 +227,14 @@ abstract class Playback(
    */
   protected abstract fun onDetachRenderer(renderer: Any?): Boolean
 
+  internal fun onRendererAttached(renderer: Any?) {
+    rendererAttachedCallback?.invoke(this, renderer)
+  }
+
+  internal fun onRendererDetached(renderer: Any?) {
+    rendererDetachedCallback?.invoke(this, renderer)
+  }
+
   internal fun onAdded() {
     "Playback#onAdded $this".logDebug()
     playbackState = STATE_ADDED
@@ -230,6 +242,8 @@ abstract class Playback(
     artworkHintListener = config.artworkHintListener
     tokenUpdateListener = config.tokenUpdateListener
     networkTypeChangeListener = config.networkTypeChangeListener
+    rendererAttachedCallback = config.rendererAttachedCallback
+    rendererDetachedCallback = config.rendererDetachedCallback
     playerParameters = networkTypeChangeListener?.onNetworkTypeChanged(manager.master.networkType)
         ?: playerParameters
     bucket.addContainer(this.container)
@@ -242,6 +256,8 @@ abstract class Playback(
     tokenUpdateListener = null
     artworkHintListener = null
     networkTypeChangeListener = null
+    rendererAttachedCallback = null
+    rendererDetachedCallback = null
     callbacks.onEach { it.onRemoved(this) }
         .clear()
     listeners.clear()