+++ kohii-core/src/main/java/kohii/v1/core/Rebinder.kt
@@ -43,6 +43,8 @@ data class Rebinder(val tag: @RawValue Any) : Parcelable {
     var artworkHintListener: ArtworkHintListener? = null
     var tokenUpdateListener: TokenUpdateListener? = null
     var networkTypeChangeListener: NetworkTypeChangeListener? = null
+    var doOnRendererAttached: RendererAvailabilityCallback? = null
+    var doOnRendererDetached: RendererAvailabilityCallback? = null
     val callbacks = mutableSetOf<Callback>()
   }
 
@@ -85,6 +87,8 @@ data class Rebinder(val tag: @RawValue Any) : Parcelable {
           it.artworkHintListener = options.artworkHintListener
           it.tokenUpdateListener = options.tokenUpdateListener
           it.networkTypeChangeListener = options.networkTypeChangeListener
+          it.doOnRendererAttached = options.doOnRendererAttached
+          it.doOnRendererDetached = options.doOnRendererDetached
           it.callbacks += options.callbacks
         }, callback
     )