+++ kohii-sample/src/main/java/kohii/v1/dev/DummyAdapter.kt
@@ -20,8 +20,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView.Adapter
 import com.google.android.exoplayer2.Player
+import com.google.android.exoplayer2.ui.PlayerView
 import kohii.v1.core.Manager
-import kohii.v1.exoplayer.DefaultControlDispatcher
+import kohii.v1.core.Playback.Controller
 import kohii.v1.exoplayer.Kohii
 import kohii.v1.sample.DemoApp
 
@@ -54,12 +55,16 @@ internal class DummyAdapter(
     kohii.setUp(DemoApp.assetVideoUri) {
       tag = "player::${holder.adapterPosition}"
       repeatMode = Player.REPEAT_MODE_ONE
-      controller = DefaultControlDispatcher(
-          manager,
-          holder.playerView,
-          kohiiCanStart = true,
-          kohiiCanPause = true
-      )
+      controller = object : Controller {
+        override fun kohiiCanStart(): Boolean = true
+        override fun kohiiCanPause(): Boolean = true
+      }
+      doOnRendererAttached = { playback, renderer ->
+        if (renderer is PlayerView) {
+          renderer.useController = true
+          renderer.setControlDispatcher(kohii.createControlDispatcher(playback))
+        }
+      }
     }
         .bind(holder.playerView)
   }