+++ kohii-sample/src/main/java/kohii/v1/sample/ui/fbook/player/BigPlayerDialog.kt
@@ -23,9 +23,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.exoplayer2.ui.AspectRatioFrameLayout
+import com.google.android.exoplayer2.ui.PlayerView
 import kohii.v1.core.Playback
+import kohii.v1.core.Playback.Controller
 import kohii.v1.core.Rebinder
-import kohii.v1.exoplayer.DefaultControlDispatcher
 import kohii.v1.exoplayer.Kohii
 import kohii.v1.sample.R
 import kohii.v1.sample.common.InfinityDialogFragment
@@ -62,6 +63,7 @@ class BigPlayerDialog : InfinityDialogFragment(),
 
   @Suppress("MemberVisibilityCanBePrivate")
   var floatPlayerController: FloatPlayerController? = null
+
   @Suppress("MemberVisibilityCanBePrivate")
   var playerCallback: Callback? = null
 
@@ -110,7 +112,7 @@ class BigPlayerDialog : InfinityDialogFragment(),
   ) {
     super.onViewCreated(view, savedInstanceState)
     kohii = Kohii[this]
-    val manager = kohii.register(this)
+    kohii.register(this)
         .addBucket(playerContainer)
 
     requireArguments().apply {
@@ -143,11 +145,19 @@ class BigPlayerDialog : InfinityDialogFragment(),
       playerCallback?.requestDismiss(this) ?: dismissAllowingStateLoss()
     }
 
-    rebinder
-        .with {
-          controller = DefaultControlDispatcher(manager, playerView)
-          callbacks += this@BigPlayerDialog
+    rebinder.with {
+      controller = object : Controller {
+        override fun kohiiCanStart(): Boolean = true
+        override fun kohiiCanPause(): Boolean = true
+      }
+      doOnRendererAttached = { playback, renderer ->
+        if (renderer is PlayerView) {
+          renderer.useController = true
+          renderer.setControlDispatcher(kohii.createControlDispatcher(playback))
         }
+      }
+      callbacks += this@BigPlayerDialog
+    }
         .bind(kohii, playerView) {
           it.addStateListener(this@BigPlayerDialog)
         }